Add files via upload
I edited the Javascript to include all brewery types with lat-long coordinates; not just brewpubs. Because of this, I also edited the HTML to note the types of locations that would show in the list. Lastly, I added line in each LI detailing what type of location each was (brewery, brewpub, etc.).